pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-intention-design"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "typer>=0.12",
    "rich>=13.7",
    "numpy>=2.3.3",
    "scikit-learn>=1.7.2"
]

[project.scripts]
idc = "idc.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
